summoner layout

+++ app/src/main/java/com/example/opgg/models/client/CLeague.kt
@@ -13,7 +13,7 @@ class CLeague(data: League?) {
     init {
         displayWinLosses =
             if(data?.wins != null && data.losses != null){
-                val rate = (data.wins.toDouble() / (data.wins + data.losses).toDouble())*100
+                val rate = (data.wins.toFloat() / (data.wins + data.losses).toFloat())*100
                 "${data.wins}승 ${data.losses}패 (${rate.toInt()}%)"
             }else{
                 ""
